Fix to #2027 - improve the way we build IsNull expressions for null semantics
Problem was that in null semantics, when we test whether expression is null or not, we assumed incorrectly that the expression is null if either of it's parameters is null.
This works for most cases (e.g. functions), but not for all.
Exapmles are COALESCE and CASE-WHEN-THEN-ELSE.

Fix is to refactor the visitor that is used for that purpose, before it was just extracting nullable expressions, now it builds the entire IsNullExpression tree.

Also fixed some minor issues with case expression - we were assuming it would always produce expression of type bool, but it's no longer the case.